Fixed bugs related with the `EnergyBar`
* `BarSegment` no longer shows the progress 'wave' if progress is 0 or 1. This was causing filled/empty segments to show the wave. This was fixed in the shader by changing the edge of the wave to not show if progress is <1% or >99%;
* `BarSegment`'s border and unfilled background are back - when swapping from the old shader to MUI_SideWave, opacity node in OutputColor was not hooked up, resulting in a black background where there was no fill in the bar instead of making it translucent;
* Fixed issue where sometimes we couldn't shoot, despite having a shot - `UShootingAttributeSet::UpdateOutOfAmmo` was not removing the 'OutOfAmmo' tag if the `CurrentEnergy` was the same as the `EnergyCostPerShot` ( if `CurrentEnergy` was at 100 and the cost of the shot was 100 and previously there was no energy for the shot, `OutOfAmmo` tag was still in place, the function would not remove it )
* Fixed issue where `BarSegment` was not visually correct when shot had a fractional point cost ( for example 12.5 when `MaxAmmo` is 8 ). Shooting was consuming a bar and removing a bit of progress from the following one. This was a problem created by using int32 instead of float to represent `FMaxAmmoChangedData::EnergyCostPerShot` and all attributes in `FEnergyAmountChangedData` (EnergyCostPerShot, NewValue, OldValue ).

+++ Source/MLlike/EnergyAmountChangedData.h
@@ -11,11 +11,11 @@ struct FEnergyAmountChangedData
 	GENERATED_BODY()
 
 	UPROPERTY()
-	int32 OldValue = -1;
+	float OldValue = -1.0f;
 
 	UPROPERTY()
-	int32 NewValue = -1;
+	float NewValue = -1.0f;
 
 	UPROPERTY()
-	int32 EnergyCostPerShot = -1;
+	float EnergyCostPerShot = -1.0f;
 };
